face_features_app/requirements.txt
deepface
insightface
onnxruntime-gpu
tf-keras
pandas
numpy
sqlalchemy
psycopg2
boto3
pyarrow
fastapi
uvicorn
dotenv
# opencv-python-headless==4.8.0.76
